Facebook Events implemented

+++ Fashion-Now/FacebookHelper.swift
@@ -17,10 +17,10 @@ class FacebookHelper {
 
     :returns: The URL of the picture
     */
-    class func urlForPictureOfUser(id facebookId: String, size avatarSize: Int?) -> NSURL? {
+    class func urlForPictureOfUser(id facebookId: String, size avatarSize: CGFloat?) -> NSURL? {
         var avatarPath = "http://graph.facebook.com/\(facebookId)/picture"
         if let unwrappedAvatarSize = avatarSize {
-            let avatarRealSize = Int(ceil(CGFloat(unwrappedAvatarSize) * UIScreen.mainScreen().scale))
+            let avatarRealSize = Int(ceil(unwrappedAvatarSize * UIScreen.mainScreen().scale))
             avatarPath += "?height=\(avatarRealSize)&width=\(avatarRealSize)"
         }
         return NSURL(string: avatarPath)
